a few refactors making it easier to add new phantoms
