+++ src/insilicoICH/ground_truth_definition/phantoms.py
@@ -4,7 +4,6 @@ module for working with phantoms
 
 from pathlib import Path
 import os
-from warnings import warn
 from collections import OrderedDict
 
 import numpy as np
@@ -12,8 +11,12 @@ import nibabel as nib
 import nrrd
 import pandas as pd
 import skimage as ski
-from dotenv import load_dotenv
 from monai.transforms import Resize, RandAffine, Affine, ResizeWithPadOrCrop
+
+from scipy.ndimage import (center_of_mass,
+                           distance_transform_edt,
+                           binary_erosion)
+
 from .utils import download_and_extract_archive
 
 from . import dicom_to_voxelized_phantom
@@ -23,9 +26,6 @@ from ..lesion_definition import (elliptical_lesion,
                                  insert_dural,
                                  warp_slice,
                                  get_perimeter)
-from scipy.ndimage import (center_of_mass,
-                           distance_transform_edt,
-                           binary_erosion)
 
 import gecatsim as xc
 
@@ -163,78 +163,7 @@ def voxelize_ground_truth(dicom_path: str | Path, phantom_path: str | Path,
     cfg.phantom.slice_range = [slice_range[0], slice_range[-1]]
     cfg.phantom.show_phantom = False
     cfg.phantom.overwrite = True
-#     cfg_file_str = fr"""
-# # Path where the DICOM images are located:
-# phantom.dicom_path = '{dicom_path}'
-# # Path where the phantom files are to be written
-# # (the last folder name will be the phantom files' base name):
-# phantom.phantom_path = '{phantom_path}'
-# phantom.materials = {list(material_threshold_dict.keys())}
-# phantom.mu_energy = 60
-# phantom.thresholds = {list(material_threshold_dict.values())}
-# phantom.slice_range = [{[slice_range[0], slice_range[-1]]}] # Range of DICOM
-# # image numbers to include. (first, last slice)
-# phantom.show_phantom = False  # Flag to turn on/off image display.
-# phantom.overwrite = True  # Flag to overwrite existing files without warning.
-# """
-
-#     dicom_to_voxel_cfg = phantom_path / 'dicom_to_voxelized.cfg'
-#     breakpoint()
-#     with open(dicom_to_voxel_cfg, 'w') as f:
-#         f.write(cfg_file_str)
-
     dicom_to_voxelized_phantom.DICOM_to_voxelized_phantom(cfg.phantom)
-    # dicom_to_voxelized_phantom.run_from_config(dicom_to_voxel_cfg)
-
-
-def load_phantom(age=38, shape=None, name='default'):
-    '''
-    Loads appropriate phantom based on age as a keyword
-
-    :param age: patient age in years, MIDA currently hard coded at 38 yrs
-    :param shape: shape of that the ground truth phantom will be interpolated
-    :param name: patient name to be saved in DICOM header
-    :param lesion_type: options include: ['IPH', 'EDH', 'SDH']
-    :param radius: lesion radius if sphere is selected
-    :param intensity: uniform intensity of the lesion (HU)
-    :param add_positioning_augmentation: bool, apply random affine to phantom
-    '''
-    load_dotenv()
-    if 'PHANTOM_DIRECTORY' in os.environ:
-        phantom_dir = Path(os.environ['PHANTOM_DIRECTORY'])
-    else:
-        phantom_dir = Path(__file__).parents[2]
-        warn(f'''
-The environment variable `PHANTOM_DIRECTORY` has not been set, this is needed
-to locate stored base phantom files for the NIHPD and MIDA head phantoms.
-
-If these phantom files cannot be located, NIHPD phantoms will be downloaded to
-your working directory: {phantom_dir}
-
-MIDA phantom files need to be downloaded manually and added to this directory,
-see `MIDA_Head_Phantom` for details.
-
-Please do one of the following:
-
-1. create a file called `.env` in this project's working directory and add:
-
-`PHANTOM_DIRECTORY=/path/to/phantoms`
-
-or
-
-2. in your terminal `export PHANTOM_DIRECTORY=/path/to_phantoms`
-''')
-
-    mida_age = 38
-    if age == mida_age:
-        phantom = MIDA_Head(phantom_dir / 'MIDA_Head_Phantom', shape=shape)
-    else:
-        phantom = NIHPD_Head(phantom_dir / 'NIHPD_Head_Phantom',
-                             age=age, shape=shape)
-
-    phantom.patient_name = name
-    phantom.age = age
-    return phantom
 
 
 class Phantom:
